requirements.txt
pip
torch
transformers
smolagents[litellm]
smolagents[transformers]
smolagents[openai]
requests
markdownify
duckduckgo-search
wikipedia
pandas
pillow
accelerate
num2words
quanto
bitsandbytes
optimum-quanto
phoenix
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-openai
sentence-transformers
chromadb
python-dotenv

